refactored test dir
